+++ src/__tests__/routes/organization.test.js
@@ -16,23 +16,26 @@ describe('Organization auth', () => {
       expect(resp.status).toBe(201);
       expect(resp.body).toHaveProperty('token');
     });
+
     test('should not sign up the user with the username which already exist', async () => {
       const resp = await request(app)
         .post(`${urlPrefix}/auth/signup`)
         .send(organization);
-      expect(resp.status).toBe(400);
+      expect(resp.status).toBe(409);
       expect(resp.body.message).toBe('Username already exist');
     });
+
     test('should not sign up the user with the email which already exist', async () => {
       const badOrg = organization;
       badOrg.username = 'new_username';
       const resp = await request(app)
         .post(`${urlPrefix}/auth/signup`)
         .send(badOrg);
-      expect(resp.status).toBe(400);
+      expect(resp.status).toBe(409);
       expect(resp.body.message).toBe('Email already exist');
     });
   });
+
   describe('Email verification', () => {
     const { email, username } = organization;
     const token = encrypt.generateToken({ email, username });
@@ -43,16 +46,19 @@ describe('Organization auth', () => {
       expect(resp.status).toBe(403);
       expect(resp.body.message).toBe('Check your email for account verification',);
     });
+
     test('should verify the account', async () => {
       const resp = await request(app).get(`${urlPrefix}/auth/verification/${token}`,);
       expect(resp.status).toBe(200);
       expect(resp.body.message).toBe('Your account has been verified successfully',);
     });
+
     test('should verify the account', async () => {
       const resp = await request(app).get(`${urlPrefix}/auth/verification/${token}`,);
       expect(resp.status).toBe(400);
       expect(resp.body.message).toBe('Your account has already been verified');
     });
+
     test('should verify the account', async () => {
       const expiredToken = jwt.sign({ email, username }, process.env.SECRET, {
         expiresIn: '0.1s',
@@ -62,6 +68,7 @@ describe('Organization auth', () => {
       expect(resp.body.message).toBe('Your verification email has expired, try to login to receive a new one',);
     });
   });
+
   describe('login with username or email verification', () => {
     test('should log the user in', async () => {
       const resp = await request(app)
@@ -70,6 +77,7 @@ describe('Organization auth', () => {
       expect(resp.status).toBe(200);
       expect(resp.body).toHaveProperty('token');
     });
+
     test('should not log the user in with wrong credentials', async () => {
       const resp = await request(app)
         .post(`${urlPrefix}/auth/login`)
@@ -78,6 +86,7 @@ describe('Organization auth', () => {
       expect(resp.body.message).toBe('The credentials you provided are incorrect',);
     });
   });
+
   afterAll(async () => {
     await User.deleteOne({ email: organization.email });
     await Organization.deleteMany({});
